+++ packages/canonical-bridge-sdk/src/layerZero/types/index.ts
@@ -46,3 +46,18 @@ export interface LayerZeroTransferConfigs {
   chains: LayerZeroChain[];
   tokens: Record<number, LayerZeroToken[]>;
 }
+
+export interface LayerZeroTokenValidateParams {
+  fromPublicClient: PublicClient;
+  toPublicClient: PublicClient;
+  bridgeAddress: `0x${string}`;
+  fromTokenAddress: `0x${string}`;
+  fromTokenSymbol: string;
+  fromTokenDecimals: number;
+  toTokenAddress: `0x${string}`;
+  toTokenSymbol: string;
+  toTokenDecimals: number;
+  toBridgeAddress: `0x${string}`;
+  dstEndpoint?: number;
+  amount: number;
+}