+++ packages/canonical-bridge-sdk/src/meson/index.ts
@@ -5,6 +5,12 @@ import {
   IMesonSendTokenInput,
 } from '@/meson/types';
 import { BaseBridgeConfig } from '@/core';
+import { isEvmAddress, isTronAddress } from '@/core/utils/address';
+import {
+  IMesonTokenList,
+  IMesonTokenValidateParams,
+} from '@/meson/types/index';
+import { VALIDATION_API_TIMEOUT } from '@/core/constants';
 
 export class Meson {
   private client?: AxiosInstance;
@@ -98,4 +104,161 @@ export class Meson {
       }
     }
   }
+
+  validateMesonToken = async ({
+    fromChainId,
+    fromChainType,
+    fromTokenSymbol,
+    fromTokenAddress,
+    fromTokenDecimals,
+    toTokenAddress,
+    toChainId,
+    toChainType,
+    toTokenSymbol,
+    toTokenDecimals,
+    amount,
+    mesonEndpoint,
+  }: IMesonTokenValidateParams) => {
+    try {
+      if (Number(amount) <= 0) {
+        console.log('Invalid token amount');
+        return false;
+      }
+      if (
+        !fromChainId ||
+        !fromTokenAddress ||
+        !fromTokenSymbol ||
+        !fromTokenDecimals ||
+        !fromChainType ||
+        !toChainId ||
+        !toTokenAddress ||
+        !toTokenDecimals ||
+        !toChainType ||
+        !toTokenSymbol ||
+        !amount ||
+        !mesonEndpoint
+      ) {
+        console.log('Missing Meson required params');
+        console.log('-- fromChainId', fromChainId);
+        console.log('-- fromTokenAddress', fromTokenAddress);
+        console.log('-- fromTokenSymbol', fromTokenSymbol);
+        console.log('-- fromTokenDecimals', fromTokenDecimals);
+        console.log('-- fromChainType', fromChainType);
+        console.log('-- toChainId', toChainId);
+        console.log('-- toTokenAddress', toTokenAddress);
+        console.log('-- toTokenDecimals', toTokenDecimals);
+        console.log('-- toChainType', toChainType);
+        console.log('-- toTokenSymbol', toTokenSymbol);
+        console.log('-- amount', amount);
+        console.log('-- mesonEndpoint', mesonEndpoint);
+        return false;
+      }
+      if (fromTokenSymbol.toLowerCase() !== toTokenSymbol.toLowerCase()) {
+        console.log('Invalid token symbol', fromTokenSymbol, toTokenSymbol);
+        return false;
+      }
+      // from token address
+      if (fromChainType === 'evm') {
+        if (!isEvmAddress(fromTokenAddress)) {
+          console.log('Invalid from token address', fromTokenAddress);
+
+          return false;
+        }
+      } else if (fromChainType === 'tron') {
+        if (!isTronAddress(fromTokenAddress)) {
+          console.log('Invalid from token address', fromTokenAddress);
+          return false;
+        }
+      } else {
+        console.log('Invalid from chain type', fromChainType);
+        return false;
+      }
+      // to token address
+      if (toChainType === 'evm') {
+        if (!isEvmAddress(toTokenAddress)) {
+          console.log('Invalid to token address', toTokenAddress);
+          return false;
+        }
+      } else if (toChainType === 'tron') {
+        if (!isTronAddress(toTokenAddress)) {
+          console.log('Invalid to token address', toTokenAddress);
+          return false;
+        }
+      } else {
+        console.log('Invalid to chain type', toChainType);
+        return false;
+      }
+      // Check token information from Meson API
+      const { data: mesonConfig } = await axios.get<{
+        result: IMesonTokenList[];
+      }>(`${mesonEndpoint}/limits`, { timeout: VALIDATION_API_TIMEOUT });
+
+      const fromHexNum =
+        fromChainType === 'tron' ? 'tron' : `0x${fromChainId?.toString(16)}`;
+      const toHexNum =
+        toChainType === 'tron' ? 'tron' : `0x${toChainId?.toString(16)}`;
+      // from token validation
+      const validFromToken = mesonConfig.result.filter((chainInfo) => {
+        const fromTokenInfo = chainInfo.tokens.filter(
+          (token) =>
+            (token?.addr?.toLowerCase() === fromTokenAddress.toLowerCase() &&
+              token.decimals === fromTokenDecimals &&
+              token.id === fromTokenSymbol.toLowerCase()) ||
+            (!token?.addr &&
+              fromTokenAddress ===
+                '0x0000000000000000000000000000000000000000' &&
+              token.decimals === fromTokenDecimals &&
+              token.id === fromTokenSymbol.toLowerCase())
+        );
+        if (!!fromTokenInfo && fromTokenInfo.length > 0) {
+          console.log('Meson from token info', fromTokenInfo);
+        }
+        return (
+          chainInfo.chainId === fromHexNum &&
+          fromTokenInfo?.length > 0 &&
+          !!fromTokenInfo
+        );
+      });
+      // to token validation
+      const validToToken = mesonConfig.result.filter((chainInfo) => {
+        const toTokenInfo = chainInfo.tokens.filter(
+          (token) =>
+            (token?.addr?.toLowerCase() === toTokenAddress.toLowerCase() &&
+              token.decimals === toTokenDecimals &&
+              token.id === toTokenSymbol.toLowerCase()) ||
+            (!token?.addr &&
+              toTokenAddress === '0x0000000000000000000000000000000000000000' &&
+              token.decimals === toTokenDecimals &&
+              token.id === toTokenSymbol.toLowerCase())
+        );
+        if (!!toTokenInfo && toTokenInfo.length > 0) {
+          console.log('Meson to token info', toTokenInfo);
+        }
+        return (
+          chainInfo.chainId === toHexNum &&
+          toTokenInfo?.length > 0 &&
+          !!toTokenInfo
+        );
+      });
+      if (validToToken?.length > 0 && validFromToken?.length > 0) {
+        console.log('Meson token info matched', validToToken, validFromToken);
+        return true;
+      }
+      console.log('Could not find Meson token');
+      console.log('-- fromChainId', fromChainId);
+      console.log('-- fromTokenAddress', fromTokenAddress);
+      console.log('-- fromTokenSymbol', fromTokenSymbol);
+      console.log('-- fromChainType', fromChainType);
+      console.log('-- toChainId', toChainId);
+      console.log('-- toTokenAddress', toTokenAddress);
+      console.log('-- toChainType', toChainType);
+      console.log('-- toTokenSymbol', toTokenSymbol);
+      console.log('-- amount', amount);
+      return false;
+      // eslint-disable-next-line
+    } catch (error: any) {
+      console.log('Meson token validation error', error);
+      return false;
+    }
+  };
 }