+++ packages/canonical-bridge-sdk/src/stargate/types/index.ts
@@ -67,3 +67,63 @@ export interface ISendTokenInput {
   receiver: `0x${string}`;
   amount: bigint;
 }
+
+export interface IStargateTokenValidateParams {
+  fromBridgeAddress: `0x${string}`;
+  toBridgeAddress: `0x${string}`;
+  fromTokenAddress: `0x${string}`;
+  fromTokenSymbol: string;
+  fromTokenDecimals: number;
+  fromChainId?: number;
+  toTokenAddress: `0x${string}`;
+  toTokenSymbol: string;
+  toTokenDecimals: number;
+  toChainId?: number;
+  amount: number;
+  toPublicClient: PublicClient;
+  fromPublicClient: PublicClient;
+  dstEndpointId: number;
+  stargateEndpoint: string;
+}
+
+export interface IStargateBridgeTokenInfo {
+  stargateType: string;
+  address: `0x${string}`;
+  token: {
+    address: `0x${string}`;
+    decimals: number;
+    symbol: string;
+  };
+  lpToken: {
+    address: `0x${string}`;
+    decimals: number;
+    symbol: string;
+  };
+  farm: {
+    stargateStaking: {
+      address: `0x${string}`;
+      rewardTokens: [
+        {
+          address: `0x${string}`;
+          decimals: number;
+          symbol: string;
+        },
+        {
+          address: `0x${string}`;
+          decimals: number;
+          symbol: string;
+        }
+      ];
+    };
+  };
+  id: string;
+  assetId: string;
+  chainKey: string;
+  chainName: string;
+  tokenMessaging: `0x${string}`;
+  sharedDecimals: number;
+}
+export interface IStargateTokenList {
+  v1: IStargateBridgeTokenInfo[];
+  v2: IStargateBridgeTokenInfo[];
+}