+++ packages/canonical-bridge-widget/src/modules/transfer/hooks/useLoadingBridgeFees.ts
@@ -168,10 +168,10 @@ export const useLoadingBridgeFees = () => {
           },
           deBridgeOpts: {
             fromChainId: fromChain.id,
-            fromTokenAddress: selectedToken.deBridge?.address as `0x${string}`,
+            fromTokenAddress: selectedToken.deBridge?.raw?.address as `0x${string}`,
             amount,
             toChainId: toChain?.id,
-            toTokenAddress: toToken?.deBridge?.address as `0x${string}`,
+            toTokenAddress: toToken?.deBridge?.raw?.address as `0x${string}`,
             accesstoken: deBridgeAccessToken,
             userAddress:
               fromChain.chainType === 'solana'
@@ -449,7 +449,7 @@ export const useLoadingBridgeFees = () => {
       toToken?.layerZero?.raw?.endpointID,
       toToken?.stargate?.raw?.endpointID,
       toToken?.meson?.raw?.id,
-      toToken?.deBridge?.address,
+      toToken?.deBridge?.raw?.address,
       max_slippage,
       tronAddress,
       deBridgeAccessToken,